+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### ✨ Enhancements
 
 - Added support for global CSS styling using the `--global-style` option in the CLI
+- Added optional attributes width and height for each drawn object
 
 ### 🐛 Bug fixes